+++ sgm/modules/diffusionmodules/video_model.py
@@ -541,7 +541,7 @@ class VideoUNet(nn.Module):
 
         if self.num_classes is not None:
             assert y.shape[0] == x.shape[0]
-            y = y.to(emb.dtype)
+            # y = y.to(emb.dtype)
             emb = emb + self.label_emb(y)
 
         h = x
